[core] move BinaryProgram to gl namespace
