+++ src/core-files.json
@@ -18,6 +18,7 @@
         "src/mbgl/geometry/line_atlas.cpp",
         "src/mbgl/gfx/attribute.cpp",
         "src/mbgl/gl/attribute.cpp",
+        "src/mbgl/gl/binary_program.cpp",
         "src/mbgl/gl/context.cpp",
         "src/mbgl/gl/debugging.cpp",
         "src/mbgl/gl/debugging_extension.cpp",
@@ -52,7 +53,6 @@
         "src/mbgl/math/log2.cpp",
         "src/mbgl/platform/gl_functions.cpp",
         "src/mbgl/programs/background_program.cpp",
-        "src/mbgl/programs/binary_program.cpp",
         "src/mbgl/programs/circle_program.cpp",
         "src/mbgl/programs/collision_box_program.cpp",
         "src/mbgl/programs/extrusion_texture_program.cpp",
@@ -520,6 +520,7 @@
         "mbgl/gfx/vertex_buffer.hpp": "src/mbgl/gfx/vertex_buffer.hpp",
         "mbgl/gfx/vertex_vector.hpp": "src/mbgl/gfx/vertex_vector.hpp",
         "mbgl/gl/attribute.hpp": "src/mbgl/gl/attribute.hpp",
+        "mbgl/gl/binary_program.hpp": "src/mbgl/gl/binary_program.hpp",
         "mbgl/gl/context.hpp": "src/mbgl/gl/context.hpp",
         "mbgl/gl/debugging.hpp": "src/mbgl/gl/debugging.hpp",
         "mbgl/gl/debugging_extension.hpp": "src/mbgl/gl/debugging_extension.hpp",
@@ -557,7 +558,6 @@
         "mbgl/map/zoom_history.hpp": "src/mbgl/map/zoom_history.hpp",
         "mbgl/programs/attributes.hpp": "src/mbgl/programs/attributes.hpp",
         "mbgl/programs/background_program.hpp": "src/mbgl/programs/background_program.hpp",
-        "mbgl/programs/binary_program.hpp": "src/mbgl/programs/binary_program.hpp",
         "mbgl/programs/circle_program.hpp": "src/mbgl/programs/circle_program.hpp",
         "mbgl/programs/clipping_mask_program.hpp": "src/mbgl/programs/clipping_mask_program.hpp",
         "mbgl/programs/collision_box_program.hpp": "src/mbgl/programs/collision_box_program.hpp",
