+++ test/programs/binary_program.test.cpp
@@ -1,16 +1,16 @@
 #include <mbgl/test/util.hpp>
 
-#include <mbgl/programs/binary_program.hpp>
+#include <mbgl/gl/binary_program.hpp>
 
 using namespace mbgl;
 
 TEST(BinaryProgram, ObtainValues) {
-    const BinaryProgram binaryProgram{ 42,
-                                       "binary code",
-                                       "identifier",
-                                       { { "a_pos", 1 }, { "a_data", 4 } },
-                                       { { "u_world", 1 }, { "u_ratio", 3 } },
-                                       { { "u_image", 0 } } };
+    const gl::BinaryProgram binaryProgram{ 42,
+                                           "binary code",
+                                           "identifier",
+                                           { { "a_pos", 1 }, { "a_data", 4 } },
+                                           { { "u_world", 1 }, { "u_ratio", 3 } },
+                                           { { "u_image", 0 } } };
 
     EXPECT_EQ(42u, binaryProgram.format());
     EXPECT_EQ("binary code", binaryProgram.code());
@@ -26,7 +26,7 @@ TEST(BinaryProgram, ObtainValues) {
 
     auto serialized = binaryProgram.serialize();
 
-    const BinaryProgram binaryProgram2(std::move(serialized));
+    const gl::BinaryProgram binaryProgram2(std::move(serialized));
 
     EXPECT_EQ(42u, binaryProgram2.format());
     EXPECT_EQ("binary code", binaryProgram2.code());
@@ -40,5 +40,5 @@ TEST(BinaryProgram, ObtainValues) {
     EXPECT_EQ(0, binaryProgram.textureLocation("u_image"));
     EXPECT_EQ(-1, binaryProgram.textureLocation("u_image2"));
 
-    EXPECT_THROW(BinaryProgram(""), std::runtime_error);
+    EXPECT_THROW(gl::BinaryProgram(""), std::runtime_error);
 }
